backend/requirements/base.txt
# Core Django packages
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
Pillow==10.1.0

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.0

# API Documentation
drf-spectacular==0.26.5

# Caching
redis==5.0.1
django-redis==5.4.0

# Task Queue
celery==5.3.4

# Email
resend==0.6.0

# File Storage
cloudinary==1.36.0
django-cloudinary-storage==0.3.0

# Security
django-ratelimit==4.1.0

# Utilities
python-decouple==3.8